+++ src/content/hooks/usePlugins.ts
@@ -1,6 +1,7 @@
 import { useMemo } from 'react'
 import remarkBreaks from 'remark-breaks'
-import { remarkCollapse, remarkHideRelativeImages } from '../remark-plugins'
+import remarkCollapse from '../remark-plugins/remark-collapse'
+import remarkHideRelativeImages from '../remark-plugins/remark-hide-relative-images'
 
 export const usePlugins = () => {
   const remarkPlugins = useMemo(() => [remarkBreaks, remarkCollapse, remarkHideRelativeImages], [])
